refactor(media): move compose ui
