+++ medialibrary/src/main/java/com/cgfay/media/recorder/AVFormatter.kt
@@ -1,17 +1,6 @@
 package com.cgfay.media.recorder
 
 import android.media.AudioFormat
-import androidx.compose.foundation.clickable
-import androidx.compose.foundation.layout.*
-import androidx.compose.material.RadioButton
-import androidx.compose.material.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.MutableState
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.unit.dp
 
 /**
  * Format definitions aligned with the native layer. This Kotlin version
@@ -45,51 +34,5 @@ object AVFormatter {
         else -> SAMPLE_FORMAT_NONE
     }
 
-    /**
-     * Jetpack Compose selector allowing users to choose an audio sample format.
-     *
-     * @param selected current selected format state
-     * @param modifier optional [Modifier]
-     * @param onFormatSelected callback when a format is chosen
-     */
-    @Composable
-    fun SampleFormatSelector(
-        selected: MutableState<Int> = remember { mutableStateOf(SAMPLE_FORMAT_16BIT) },
-        modifier: Modifier = Modifier,
-        onFormatSelected: (Int) -> Unit = {},
-    ) {
-        val formats = listOf(SAMPLE_FORMAT_8BIT, SAMPLE_FORMAT_16BIT, SAMPLE_FORMAT_FLOAT)
-        Column(modifier = modifier) {
-            formats.forEach { format ->
-                Row(
-                    verticalAlignment = Alignment.CenterVertically,
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .clickable {
-                            selected.value = format
-                            onFormatSelected(format)
-                        }
-                        .padding(vertical = 4.dp)
-                ) {
-                    RadioButton(
-                        selected = selected.value == format,
-                        onClick = {
-                            selected.value = format
-                            onFormatSelected(format)
-                        }
-                    )
-                    Spacer(modifier = Modifier.width(8.dp))
-                    Text(text = formatLabel(format))
-                }
-            }
-        }
-    }
-
-    private fun formatLabel(format: Int): String = when (format) {
-        SAMPLE_FORMAT_8BIT -> "8-bit PCM"
-        SAMPLE_FORMAT_16BIT -> "16-bit PCM"
-        SAMPLE_FORMAT_FLOAT -> "Float PCM"
-        else -> "Unknown"
-    }
 }
