+++ medialibrary/src/main/java/com/cgfay/media/recorder/MediaType.kt
@@ -1,11 +1,10 @@
 package com.cgfay.media.recorder
 
-import androidx.compose.runtime.Stable
 
 /**
  * Type of media being recorded.
  */
-@Stable
+enum class MediaType {
 enum class MediaType {
     AUDIO,
     VIDEO
